No need to load GO model every time you run server

+++ src/ai/model.py
@@ -243,8 +243,10 @@ class GeneOntologyPrediction(BaseModel):
         local_path = dirname(os.path.abspath(__file__)) + "/custom_models/models/gene_ontology/go_model_150M.pth"
 
         response = requests.get(model_url)
-        with open(local_path, 'wb') as f:
-            f.write(response.content)
+
+        if not os.path.exists(local_path):
+            with open(local_path, 'wb') as f:
+                f.write(response.content)
 
         self.model.load_state_dict(torch.load(local_path, map_location=self.device))
